chore(coinjoin): split BalanceSection to make it more readable
